Problem: new project structuring dictates a crate at root level
Solution: remove gaslighter as it did too many things and
fill out `blockchain_geth` and `regression_test`.
(note `trackertests` is now named `regression_test`)
