+++ blockchain_geth/Cargo.toml
@@ -1,6 +1,13 @@
 [package]
 name = "blockchain_geth"
 version = "0.1.0"
+license = "Apache-2.0"
 authors = ["Stewart Mackenzie <setori88@gmail.com>", "Wei Tang <hi@that.world>"]
+description = "blockchain_geth - provides functionality for interacting with geth's blockchain."
+
+[lib]
+name = "blockchain_geth"
 
 [dependencies]
+clap = "2.22"
+serde_json = "0.9"
