+++ regression_test/Cargo.toml
@@ -1,14 +1,15 @@
 [package]
-name = "gaslighter"
+name = "regression_test"
 version = "0.1.0"
 license = "Apache-2.0"
 authors = ["Stewart Mackenzie <setori88@gmail.com>", "Wei Tang <hi@that.world>"]
-description = "SputnikVM - a Virtual Machine tester"
+description = "regression_test - performs a regression test of the entire ethereum classic blockchain."
 
 [[bin]]
-name = "gaslighter"
+name = "regression_test"
 
 [dependencies]
 sputnikvm = { path = '../sputnikvm' }
+blockchain_geth = { path = '../blockchain_geth' }
 clap = "2.22"
 serde_json = "0.9"
