--- trackertests/src/lib.rs
@@ -1,6 +1,0 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-    }
-}
